refactor(cli): split the security gauntlet out of ui.rs
`ui.rs` was 1157 lines against a 1200-line ceiling — forty-three lines of
room, which means the next feature in it fails the gate after the work is
done rather than before it starts. The repo's rule is to split before the
limit; this is that.

`ui_gauntlet.rs` takes the whole gauntlet: ten real attacks on the
running kernel plus the baseline that proves the harness itself works,
all of it in memory against the live crates. The cut is there because the
dependency runs one way — the gauntlet uses nothing from the HTTP
surface, and the HTTP surface uses one function from it. 647 and 527
lines, both well clear of the ceiling.

A pure move: the same 90 workspace tests and 7 test binaries pass
unchanged, and `cargo fix` removed the imports each file stopped needing
rather than my guessing at them.

Verified through the running route rather than only by compiling, because
a move that builds can still have stopped doing anything: `POST
/api/gauntlet` returns 11 results, all passing, one of them the baseline.

That count corrected a stale comment. The banner said "the seven
attacks"; there are ten, and the interface has said ten all along. The
comment was the thing that had drifted.
